completed routing and seperated customer view and admin view and header footer for customer
